fix: wrong migration deleted

+++ database/migrations/2026_05_25_191227_change_topics_column_on_questions_table.php
@@ -10,21 +10,34 @@ return new class extends Migration
      * Run the migrations.
      */
     public function up(): void
-    {
+{
+    if (Schema::hasColumn('questions', 'topic')) {
         Schema::table('questions', function (Blueprint $table) {
             $table->dropColumn('topic');
-            $table->json('topics')->nullable();
         });
     }
 
+    if (! Schema::hasColumn('questions', 'topics')) {
+        Schema::table('questions', function (Blueprint $table) {
+            $table->json('topics')->nullable();
+        });
+    }
+}
     /**
      * Reverse the migrations.
      */
     public function down(): void
-    {
+{
+    if (Schema::hasColumn('questions', 'topics')) {
         Schema::table('questions', function (Blueprint $table) {
             $table->dropColumn('topics');
+        });
+    }
+
+    if (! Schema::hasColumn('questions', 'topic')) {
+        Schema::table('questions', function (Blueprint $table) {
             $table->string('topic')->nullable();
         });
     }
+}
 };
